PointSet.Point() performance test (successful).
For Grid.AllClusters() to perform optimally (linear wrt the number of
point in the grid), PointSet.Point() must be an O(1) operation with
respect to the size of the PointSet. Test verifies that this is true.

Additionally, found that trying to initialize a full PointSet much
bigger than 2048 * 2048 causes an out-of memory error (on a VM with
2 GB of RAM).  What does this say about the memory efficiency of
PointSet?